legacy/requirements.txt
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
msoffcrypto-tool>=5.0.0
